# backend/fastapi/app/routers/admission.py
# generated by fastapi-codegen:
#   filename:  openapi.yaml
#   timestamp: 2026-06-30T16:24:15+00:00

from __future__ import annotations

from fastapi import APIRouter

from ..dependencies import *

router = APIRouter(tags=['Admission'])


@router.post('/admissionForm', response_model=AdmissionForm, tags=['Admission'])
def post_admission_form(body: AdmissionFormCreate) -> AdmissionForm:
    """
    Submit Admission Form
    """
    pass


@router.get('/admissionForm/me', response_model=List[AdmissionForm], tags=['Admission'])
def get_admission_form_me() -> List[AdmissionForm]:
    """
    My Admission Forms
    """
    pass


@router.post(
    '/admissionForm/{id}/approved', response_model=StudentFull, tags=['Admission']
)
def post_admission_form_id_approved(id: int) -> StudentFull:
    """
    Approve Form
    """
    pass


@router.post('/admissionForm/{id}/declined', response_model=None, tags=['Admission'])
def post_admission_form_id_declined(id: int) -> None:
    """
    Decline Form
    """
    pass


@router.get(
    '/admissionFormList', response_model=List[AdmissionForm], tags=['Admission']
)
def get_admission_form_list(
    status: Optional[AdmissionStatus] = None,
) -> List[AdmissionForm]:
    """
    Get Submitted Forms (Admin)
    """
    pass