# backend/fastapi/app/routers/courses.py
# generated by fastapi-codegen:
#   filename:  openapi.yaml
#   timestamp: 2026-06-30T16:24:15+00:00

from __future__ import annotations

from fastapi import APIRouter

from ..dependencies import *

router = APIRouter(tags=['Courses'])


@router.post('/course', response_model=Course, tags=['Courses'])
def post_course(body: CourseCreate) -> Course:
    """
    Create Course
    """
    pass


@router.put('/course/{id}', response_model=Course, tags=['Courses'])
def put_course_id(id: int, body: CourseCreate = ...) -> Course:
    """
    Update Course
    """
    pass


@router.delete('/course/{id}', response_model=None, tags=['Courses'])
def delete_course_id(id: int) -> None:
    """
    Delete Course
    """
    pass


@router.get('/courseList', response_model=List[Course], tags=['Courses'])
def get_course_list() -> List[Course]:
    """
    Get list of all courses
    """
    pass