# backend/fastapi/app/routers/payments.py
# generated by fastapi-codegen:
#   filename:  openapi.yaml
#   timestamp: 2026-06-30T16:24:15+00:00

from __future__ import annotations

from fastapi import APIRouter

from ..dependencies import *

router = APIRouter(tags=['Payments'])


@router.post('/student/{id}/payment', response_model=Payment, tags=['Payments'])
def post_student_id_payment(id: int, body: PaymentCreate = ...) -> Payment:
    """
    Record a Payment
    """
    pass


@router.put(
    '/student/{id}/payment/{payment_id}', response_model=Payment, tags=['Payments']
)
def put_student_id_payment_payment_id(
    id: int, payment_id: int = ..., body: PaymentCreate = ...
) -> Payment:
    """
    Correct a Payment (supersede)
    """
    pass


@router.delete(
    '/student/{id}/payment/{payment_id}', response_model=None, tags=['Payments']
)
def delete_student_id_payment_payment_id(id: int, payment_id: int = ...) -> None:
    """
    Delete a Payment (hard delete)
    """
    pass


@router.get(
    '/student/{id}/paymentList', response_model=List[Payment], tags=['Payments']
)
def get_student_id_payment_list(id: int) -> List[Payment]:
    """
    Get Payment History for a Student
    """
    pass