# backend/fastapi/app/routers/students.py
# generated by fastapi-codegen:
#   filename:  openapi.yaml
#   timestamp: 2026-06-30T16:24:15+00:00

from __future__ import annotations

from fastapi import APIRouter

from ..dependencies import *

router = APIRouter(tags=['Students'])


@router.post('/student', response_model=StudentFull, tags=['Students'])
def post_student(body: StudentCreate) -> StudentFull:
    """
    Create Student (manual, admin-only)
    """
    pass


@router.get(
    '/student/{id}', response_model=Union[StudentFull, StudentBasic], tags=['Students']
)
def get_student_id(id: int) -> Union[StudentFull, StudentBasic]:
    """
    Get Student
    """
    pass


@router.put('/student/{id}', response_model=StudentFull, tags=['Students'])
def put_student_id(id: int, body: StudentUpdate = ...) -> StudentFull:
    """
    Update Student
    """
    pass


@router.delete('/student/{id}', response_model=None, tags=['Students'])
def delete_student_id(id: int) -> None:
    """
    Delete Student
    """
    pass


@router.get(
    '/studentList',
    response_model=List[Union[StudentFull, StudentBasic]],
    tags=['Students'],
)
def get_student_list(
    department: Optional[Department] = None,
    admission_type: Optional[AdmissionType] = None,
    learning_mode: Optional[LearningMode] = None,
    fee_type: Optional[FeeType] = None,
    batch: Optional[Batch] = None,
    fees_min: Optional[float] = None,
    fees_max: Optional[float] = None,
    joined_after: Optional[date] = None,
    joined_before: Optional[date] = None,
    search: Optional[str] = None,
) -> List[Union[StudentFull, StudentBasic]]:
    """
    Get Students List
    """
    pass